+++ server/aws-lsp-codewhisperer/src/shared/amazonQServiceManager/AmazonQTokenServiceManager.ts
@@ -27,11 +27,7 @@ import {
     QServiceManagerFeatures,
 } from './BaseAmazonQServiceManager'
 import { AWS_Q_ENDPOINTS, Q_CONFIGURATION_SECTION } from '../constants'
-import {
-    AmazonQDeveloperProfile,
-    getListAllAvailableProfilesHandler,
-    signalsAWSQDeveloperProfilesEnabled,
-} from './qDeveloperProfiles'
+import { AmazonQDeveloperProfile, signalsAWSQDeveloperProfilesEnabled } from './qDeveloperProfiles'
 import { isStringOrNull } from '../utils'
 import { getAmazonQRegionAndEndpoint } from './configurationUtils'
 import { getUserAgent } from '../telemetryUtils'
@@ -87,6 +83,11 @@ export class AmazonQTokenServiceManager extends BaseAmazonQServiceManager<
 
     private constructor(features: QServiceManagerFeatures) {
         super(features)
+    }
+
+    // @VisibleForTesting, please DO NOT use in production
+    setState(state: 'PENDING_CONNECTION' | 'PENDING_Q_PROFILE' | 'PENDING_Q_PROFILE_UPDATE' | 'INITIALIZED') {
+        this.state = state
     }
 
     public static initInstance(features: QServiceManagerFeatures): AmazonQTokenServiceManager {
@@ -151,6 +152,7 @@ export class AmazonQTokenServiceManager extends BaseAmazonQServiceManager<
         try {
             if (params.section === Q_CONFIGURATION_SECTION && params.settings.profileArn !== undefined) {
                 const profileArn = params.settings.profileArn
+                const region = params.settings.region
 
                 if (!isStringOrNull(profileArn)) {
                     throw new Error('Expected params.settings.profileArn to be of either type string or null')
@@ -318,21 +320,21 @@ export class AmazonQTokenServiceManager extends BaseAmazonQServiceManager<
         }
 
         const parsedArn = parse(newProfileArn)
-        const endpoint = AWS_Q_ENDPOINTS.get(parsedArn.region)
+        const region = parsedArn.region
+        const endpoint = AWS_Q_ENDPOINTS.get(region)
         if (!endpoint) {
             throw new Error('Requested profileArn region is not supported')
         }
 
-        const profiles = await getListAllAvailableProfilesHandler(this.serviceFactory)({
-            connectionType: 'identityCenter',
-            logging: this.logging,
-            token: token,
-            endpoints: new Map([[parsedArn.region, endpoint]]),
-        })
-
-        this.handleTokenCancellationRequest(token)
-
-        const newProfile = profiles.find(el => el.arn === newProfileArn)
+        // Hack to inject a dummy profile name as it's not used by client IDE for now, if client IDE starts consuming name field then we should also pass both profile name and arn from the IDE
+        // When service is ready to take more tps, revert https://github.com/aws/language-servers/pull/1329 to add profile validation
+        const newProfile: AmazonQDeveloperProfile = {
+            arn: newProfileArn,
+            name: 'Client provided profile',
+            identityDetails: {
+                region: parsedArn.region,
+            },
+        }
 
         if (!newProfile || !newProfile.identityDetails?.region) {
             this.log(`Amazon Q Profile ${newProfileArn} is not valid`)