+++ server/aws-lsp-codewhisperer/src/shared/telemetry/types.ts
@@ -75,6 +75,12 @@ export interface CodeWhispererCodePercentageEvent {
     successCount: number
 }
 
+export interface UserWrittenPercentageEvent {
+    codewhispererLanguage: string
+    userWrittenCodeCharacterCount: number
+    userWrittenCodeLineCount: number
+}
+
 export interface CodeWhispererUserDecisionEvent {
     codewhispererRequestId?: string
     codewhispererSessionId?: string
@@ -172,6 +178,10 @@ export enum ChatTelemetryEventName {
     ModifyCode = 'amazonq_modifyCode',
     ToolUseSuggested = 'amazonq_toolUseSuggested',
     InteractWithAgenticChat = 'amazonq_interactWithAgenticChat',
+    LoadHistory = 'amazonq_loadHistory',
+    ChatHistoryAction = 'amazonq_performChatHistoryAction',
+    ExportTab = 'amazonq_exportTab',
+    UiClick = 'ui_click',
 }
 
 export interface ChatTelemetryEventMap {
@@ -187,6 +197,10 @@ export interface ChatTelemetryEventMap {
     [ChatTelemetryEventName.ModifyCode]: ModifyCodeEvent
     [ChatTelemetryEventName.ToolUseSuggested]: ToolUseSuggestedEvent
     [ChatTelemetryEventName.InteractWithAgenticChat]: InteractWithAgenticChatEvent
+    [ChatTelemetryEventName.LoadHistory]: LoadHistoryEvent
+    [ChatTelemetryEventName.ChatHistoryAction]: ChatHistoryActionEvent
+    [ChatTelemetryEventName.ExportTab]: ExportTabEvent
+    [ChatTelemetryEventName.UiClick]: UiClickEvent
 }
 
 export type ToolUseSuggestedEvent = {
@@ -236,6 +250,7 @@ export type AddMessageEvent = {
     cwsprChatConversationType: ChatConversationType
     codewhispererCustomizationArn?: string
     languageServerVersion?: string
+    requestIds?: string[]
 
     // context related metrics
     cwsprChatHasContextList?: boolean
@@ -269,6 +284,33 @@ export type ExitFocusConversationEvent = {
     cwsprChatConversationId: string
 }
 
+export type UiClickEvent = {
+    elementId: string
+}
+
+export type LoadHistoryEvent = {
+    amazonqTimeToLoadHistory: number
+    amazonqHistoryFileSize: number
+    openTabCount: number
+    result: Result
+    languageServerVersion?: string
+}
+
+export type ChatHistoryActionEvent = {
+    action: ChatHistoryActionType
+    result: Result
+    languageServerVersion?: string
+    filenameExt?: string
+    amazonqTimeToSearchHistory?: number
+    amazonqHistoryFileSize?: number
+}
+
+export type ExportTabEvent = {
+    filenameExt: string
+    result: Result
+    languageServerVersion?: string
+}
+
 export enum ChatInteractionType {
     InsertAtCursor = 'insertAtCursor',
     CopySnippet = 'copySnippet',
@@ -279,6 +321,13 @@ export enum ChatInteractionType {
     Upvote = 'upvote',
     Downvote = 'downvote',
     ClickBodyLink = 'clickBodyLink',
+}
+
+export enum ChatHistoryActionType {
+    Search = 'search',
+    Export = 'export',
+    Open = 'open',
+    Delete = 'delete',
 }
 
 export type ChatConversationType = 'Chat' | 'Assign' | 'Transform' | 'AgenticChat' | 'AgenticChatWithToolUse'